Translations for presets optional
In this commit, I removed the translations for presets from UIText.lua, so that they are now only bound by Translation.lua.

Since in this case the author of the preset can decide for himself whether he wants to enable translations, presets should now be usable as pure data packages again.

If no translation can be found for a preset, it simply loads the data set by the author. Only an update of the translation is required and not the entire mod.