4.9.0xl
Restructured, rewritten, and despaghettified code:

- greatly improved modularity
- clearer structure of the code with debloated init.lua
- developed Config.lua with new methods and other general methods moved from other modules
- developed Translation system (introduced by Neclaex) to match changes in other modules (now it's more universal and can translate any table)
- presets IDs are alphanumeric after all

The base functionality remains the same as in the previous version. For now, it's with the empty Settings.lua module, as I lost a lot of time having some problems with the new version of it.

I'll implement a less sophisticated solution and add it along with other new features planned for 4.9.0 in upcoming alphas